nvme.h: unclutter global header file
This header files contains random function and definition. Move them to
a new home where they are group with other matching functions and
definition.
