+++ shared/time-util.h
@@ -4,6 +4,7 @@
  */
 #pragma once
 
+#include <sys/time.h>
 #include <time.h>
 
 /*
@@ -11,3 +12,6 @@
  * ts_buf, which must be at least 32 bytes. Return: 0.
  */
 int shr_format_ts(time_t time_ms, char *ts_buf);
+
+unsigned long long shr_elapsed_utime(struct timeval start_time,
+				      struct timeval end_time);
