+++ src/plugin.h
@@ -2,6 +2,9 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
+
+#include "nvme-json.h"
 
 struct program {
 	const char *name;
@@ -32,3 +35,8 @@ struct command {
 
 void general_help(struct plugin *plugin, char *str);
 int handle_plugin(int argc, char **argv, struct plugin *plugin);
+
+void register_extension(struct plugin *plugin);
+
+int __id_ctrl(int argc, char **argv, struct command *acmd,
+	struct plugin *plugin, void (*vs)(uint8_t *vs, struct json_object *root));
